move RasterRDD minMax tests
